+++ src/components/switch/switch.tsx
@@ -6,7 +6,7 @@ const rootClassNames = classNames(
     'group peer w-10 cursor-default rounded-[40px] border border-neutral-200 bg-neutral-0 p-1', // Default
     'data-[state=checked]:border-primary-400 data-[state=checked]:shadow-primary-md', // State is checked
     'focus:outline-none focus-visible:ring focus-visible:ring-primary focus-visible:ring-offset', // Focus
-    'disabled:cursor-not-allowed disabled:bg-neutral-100 disabled:data-[state=checked]:border-neutral-200 disabled:data-[state=checked]:shadow-none', // Disabled
+    'disabled:bg-neutral-100 disabled:data-[state=checked]:border-neutral-200 disabled:data-[state=checked]:shadow-none', // Disabled
 );
 
 const thumbClassNames = classNames(